Refactor patch code
Split the patch code into smaller modules to make it easier to work
with. Current commit should olny be moving of code, not changing
anything.
